accessibility-analyzer-backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
firecrawl-py
pydantic
httpx
